Added custom, protobuf-build of quiche
This includes a ton of changes... here is a quick overview of things to
know:
- many project/cargo files were modified to include our custom cache
  files (custom_cache.rs, http_record.rs), along with including some
extra things like the protobuf version, the compiled protobuf structure,
etc.
- the `docker-protobuf-build` target was added to the Makefile to build
  our Docker images
- many scripts were added that can be leveraged by the testbed to
  spinup/teardown components of the quiche server. these include things
like `start-tcpdump.sh`--which helps keep a pcap artifact for debugging, `copy-protobuf-files.hs`--which moves the protobuf files into the right place, etc. these ideally will be modified so they are not so clunky in the future