Added resource priority logging via JSON
implemented the ability to log the priority info of resources via the
`--priority-json` flag for quiche-server.

things of note:
- this does file I/O after every handle_requests call. this is because
the server has no graceful shutdown. I don't see any problems with this
in its current state but as the number of resources drastically
increases this may be an issue
- if no priority-json arg is supplied, /priorities.json is the default
- start-quiche.json has been modified to supply this argument from the
runtime-dir index