Added cache_key to the priority log message
This involved reworking a bit of how the cache_key is created/passed
around. Previously this was just created internally for the
build_h3_response function, but we needed access to it outside of that
function. Instead, I have a (default-ish) struct created and passed into
build_h3_response, where it is then modified. Remains to be tested